feat(Agent): added changes to the constructor to be able to inject a custom MultitypeQueue object. uses MockMultitypeQueue in Agent's UTs
