+++ src/agent/tests/message_queue_utils_test.cpp
@@ -3,53 +3,14 @@
 #include <command_entry.hpp>
 #include <message.hpp>
 
-#include <imultitype_queue.hpp>
+#include <mock_multitype_queue.hpp>
 
 #include <boost/asio.hpp>
-#include <gmock/gmock.h>
 #include <gtest/gtest.h>
 #include <nlohmann/json.hpp>
 
 const nlohmann::json BASE_DATA_CONTENT =
     R"({"document_id":"112233", "action":{"name":"command_test","args":{"parameters":["parameters_test"]}}})"_json;
-
-class MockMultiTypeQueue : public IMultiTypeQueue
-{
-public:
-    MOCK_METHOD(int, push, (Message message, bool shouldWait), (override));
-    MOCK_METHOD(boost::asio::awaitable<int>, pushAwaitable, (Message message), (override));
-    MOCK_METHOD(int, push, (std::vector<Message> messages), (override));
-    MOCK_METHOD(Message,
-                getNext,
-                (MessageType type, const std::string module, const std::string moduleType),
-                (override));
-    MOCK_METHOD(boost::asio::awaitable<std::vector<Message>>,
-                getNextBytesAwaitable,
-                (MessageType type, const size_t, const std::string moduleName, const std::string moduleType),
-                (override));
-    MOCK_METHOD(std::vector<Message>,
-                getNextBytes,
-                (MessageType type, const size_t, const std::string moduleName, const std::string moduleType),
-                (override));
-    MOCK_METHOD(bool, pop, (MessageType type, const std::string moduleName, const std::string moduleType), (override));
-    MOCK_METHOD(int,
-                popN,
-                (MessageType type, int messageQuantity, const std::string moduleName, const std::string moduleType),
-                (override));
-    MOCK_METHOD(bool,
-                isEmpty,
-                (MessageType type, const std::string moduleName, const std::string moduleType),
-                (override));
-    MOCK_METHOD(bool,
-                isFull,
-                (MessageType type, const std::string moduleName, const std::string moduleType),
-                (override));
-    MOCK_METHOD(int,
-                storedItems,
-                (MessageType type, const std::string moduleName, const std::string moduleType),
-                (override));
-    MOCK_METHOD(size_t, sizePerType, (MessageType type), (override));
-};
 
 class MessageQueueUtilsTest : public ::testing::Test
 {
